feat: moved test helpers to belong to the thing they're helping to test
