+++ storage/src/storage/property/mod.rs
@@ -1,11 +1,11 @@
-mod base_store;
-pub use base_store::*;
+pub mod base_store;
+pub use base_store::BaseStore;
 
-mod recall_by_id;
-pub use recall_by_id::*;
+pub mod recall_by_id;
+pub use recall_by_id::RecallById;
 
-mod recall_by_company;
-pub use recall_by_company::*;
+pub mod recall_by_company;
+pub use recall_by_company::RecallByCompany;
 
-mod recall_by_name;
-pub use recall_by_name::*;
+pub mod recall_by_name;
+pub use recall_by_name::RecallByName;
