+++ storage/src/storage/property/recall_by_company.rs
@@ -7,3 +7,24 @@ where
 {
     async fn recall_by_company<I: HasId>(&self, company_id: &I) -> Result<Vec<T>>;
 }
+
+#[cfg(test)]
+pub mod test_helper {
+    macro_rules! test_recall_by_company {
+        ($storage:ident, $storable:ident) => {
+            paste! {
+                #[tokio::test]
+                async fn [< test_recall_by_company_ $storage:snake _with_ $storable:snake >] () {
+                    let mut test_subject = $storage::new_test().await.expect("Could not create storage");
+                    let storable = $storable::new_test().await.expect("Could not create storable");
+                    test_subject.store(storable.clone()).await.expect("Could not store storable in storage");
+                    let recalled_storable = test_subject.recall_by_company(&storable.get_company_id()).await.expect("Could not recall storable from storage by id");
+                    assert_eq!(recalled_storable.len(), 1);
+                    assert!(recalled_storable.contains(&storable));
+                }
+            }
+        };
+    }
+
+    pub(crate) use test_recall_by_company;
+}
